refactor: optimizar importaciones y eliminar código no utilizado en componentes de tablas
