+++ src/app/dashboard/admin/quotation-groups/_components/quotation-group.column.tsx
@@ -1,6 +1,6 @@
 import { ColumnDef } from "@tanstack/react-table";
 
-import { DataTableColumnHeader } from "@/shared/components/data-table/DataTableColumnHeaderProps";
+import { DataTableColumnHeader } from "@/shared/components/data-table/data-table-column-header";
 import { Badge } from "@/shared/components/ui/badge";
 import { QuotationGroupResponse } from "../_types/quotation-groups.types";
 import QuotationGroupActions from "./QuotationGroupActions";
